requirements.txt
# ── Core AI / NLP ──
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ── OCR / Computer Vision ──
easyocr>=1.7.0
Pillow>=10.0.0
pdf2image>=1.16.0
opencv-python>=4.8.0

# ── LLM API Clients ──
requests>=2.31.0
httpx>=0.27.0

# ── Dataset handling ──
pandas>=2.1.0
datasets>=2.18.0

# ── Configuration ──
python-dotenv>=1.0.0

# ── UI ──
gradio>=4.25.0

# ── Evaluation metrics ──
scipy>=1.11.0
